Added generation of documentation for data object types used in JAX-RS resource methods when DTO serialization is done using a Pojo style (i.e. Jackson POJO -> JSON serialization).
Data object types that are not external, but are resolvable in the current Javadoc source path (which can be extended using standard javadoc include mechanisms) are documented at a field level, and linked from the method details.

A new top level Data objects page is generated and added to the menu, listing all data objects resolved during resource documentation.